Data.List.End

Documentation

insertBy :: Fold a e -> (a -> a -> Ordering) -> a -> EndList e' a -> EndList e a

Insert with a given Ordering and Fold to update the end

insertByM :: Monad m => FoldM m a e -> (a -> a -> Ordering) -> a -> EndList e' a -> m (EndList e a)

Insert with a given Ordering and FoldM to update the end

maximumBy :: (a -> a -> Ordering) -> EndList e a -> a

Maximum using the given ordering

minimumBy :: (a -> a -> Ordering) -> EndList e a -> a

Minimum using the given ordering

genericLength :: Num i => (e -> i) -> EndList e a -> i

Generic length, using the supplied function to calculate the length of the end.

strictGenericLength :: Num i => (e -> i) -> EndList e a -> i

Strict version of genericLength

genericTake :: Integral i => Fold a e -> i -> EndList e' a -> EndList e a

Take some number of elements and use the given Fold to calculate the new end

genericTakeM :: (Integral i, Monad m) => FoldM m a e -> i -> EndList e' a -> m (EndList e a)

genericTake, where a FoldM is provided to update the end

genericDrop :: Integral i => i -> EndList e a -> EndList e a

Drop the given number of elements from the front of an EndList or none if negative.

genericSplitAt :: Integral i => Fold a e' -> i -> EndList e a -> (EndList e' a, EndList e a)

Split at the given index, using the given Fold to provide the end of the prefix, or split at 0 if negative.

genericSplitAtM :: (Integral i, Monad m) => FoldM m a e' -> i -> EndList e a -> m (EndList e' a, EndList e a)

genericSplitAt, where a FoldM is provided to calculate the end of the prefix

genericIndex :: Integral i => EndList e a -> i -> a

The element of an EndList at the given index. Throws an error if given a negative argument

data EndList e a

A list of elements of type a where the end of the list contains a single element of type e

Constructors

End e

(:.) a (EndList e a)

fromListWithEnd :: e -> [a] -> EndList e a

Convert a list and an end value to an EndList

fromListWithEndF :: (a -> e -> e) -> e -> [a] -> EndList e a

Convert a list and an end value calculated through a left scan to an EndList

fromListWithEndA :: Applicative f => (a -> f e -> f e) -> f e -> [a] -> f (EndList e a)

fromListWithEndF with an Applicative state

ehead :: EndList e a -> a

The first element of an EndList, lazily matched

etail :: EndList e a -> EndList e a

The tail of an EndList, lazily matched

econcat :: Semigroup e => EndList e (EndList e a) -> EndList e a

Concatenate EndLists, combining their ends with (<>)

econcatWith :: (e -> e -> e) -> EndList e (EndList e a) -> EndList e a

Concatenate EndLists, combining their ends with the given method

endListToFree :: EndList e a -> Free ((,) a) e

Convert an EndList to a Free representation

freeToEndList :: Free ((,) a) e -> EndList e a

Convert a Free representation to an EndList

endListToFreeM :: Monad m => EndList e a -> FreeT ((,) a) m e

Convert an EndList to a Monadic FreeT representation

freeToEndListM :: Monad m => FreeT ((,) a) m e -> m (EndList e a)

Convert a Monadic FreeT representation to an EndList

resetEndWith :: EndList e' a -> Fold a e -> EndList e a

Use a Fold to replace the end of an EndList

resetEndWithM :: Monad m => EndList e' a -> FoldM m a e -> m (EndList e a)

Use a FoldM to replace the end of an EndList

spanEnd :: (a -> Bool) -> [a] -> EndList [a] a

span, where the prefix that satisfies the predicate is the non-end part of the EndList and the rest of the input is stored in the End.

For example:

 λ> spanEnd (< 4) [1..6]
 1 :. 2 :. 3 :. End [4,5,6]

newtype PartList a b

Example application of EndList: a list that's all b's, then all a's, then all b's, ..

Constructors

PartList

Fields

runPartList :: EndList (Maybe (PartList b a)) b

partEither :: [Either a b] -> Either (PartList b a) (PartList a b)

Convert a list of Either's to a Either PartList possibility

enumEither :: Integral t => t -> [Either () ()]

Left for even bits and Right for odd:

 λ> enumEither 1265
 [Right (),Left (),Left (),Left (),Right (),Right (),Right (),Right (),Left (),Left (),Right ()]--

enumEithers :: (Enum b, Num b, Integral t) => t -> [Either b b]

Convert an Integral input to a list of Eithers whose values are [0..] and are Left or Right as in enumEither

 λ> enumEithers 1265
 [Right 0,Left 1,Left 2,Left 3,Right 4,Right 5,Right 6,Right 7,Left 8,Left 9,Right 10]

 λ> partEither $ enumEithers 1265
 Right
   (PartList
      { runPartList =
          0 :.
          End
            Just
            (PartList
               { runPartList =
                   1 :. 2 :. 3 :.
                   End
                     Just
                     (PartList
                        { runPartList =
                            4 :. 5 :. 6 :. 7 :.
                            End
                              Just
                              (PartList
                                 { runPartList =
                                     8 :. 9 :.
                                     End
                                       Just
                                       (PartList
                                          {runPartList = 10 :. End Nothing})
                                 })
                        })
               })
      })